requirements.txt
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
python-docx==0.8.11
joblib==1.3.2
pillow==12.1.0